LPD-26752 Move test setup into a new util files
