+++ modules/test/playwright/tests/blogs-web/blogEntry.spec.ts
@@ -10,12 +10,10 @@ import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
 import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
 import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
-import {clickAndExpectToBeVisible} from '../../utils/clickAndExpectToBeVisible';
 import getRandomString from '../../utils/getRandomString';
 import {waitForSuccessAlert} from '../../utils/waitForSuccessAlert';
-import getPageDefinition from '../layout-content-page-editor-web/utils/getPageDefinition';
-import getWidgetDefinition from '../layout-content-page-editor-web/utils/getWidgetDefinition';
 import {blogsPagesTest} from './fixtures/blogsPagesTest';
+import {friendlyURLCategoriesSetup} from './utils/friendlyURLCategoriesSetup';
 
 const test = mergeTests(
 	apiHelpersTest,
@@ -77,94 +75,15 @@ test('LPD-26752 Select categories for the custom friendly URL', async ({
 	const vocabularyName = getRandomString();
 	const friendlyUrlCategories = ['category-1', 'category-2', 'category-3'];
 
-	const {id: vocabularyId} =
-		await apiHelpers.headlessAdminTaxonomy.createVocabulary({
-			name: vocabularyName,
-			siteId: site.id,
-		});
-
-	for (const categoryName of friendlyUrlCategories) {
-		await apiHelpers.headlessAdminTaxonomy.createCategory({
-			name: categoryName,
-			vocabularyId,
-		});
-	}
-
-	await displayPageTemplatesPage.goto(site.friendlyUrlPath);
-
-	const displayPageTemplateName = getRandomString();
-
-	await displayPageTemplatesPage.publishNewTemplate({
-		contentType: 'Blogs Entry',
-		name: displayPageTemplateName,
-	});
-
-	await displayPageTemplatesPage.markAsDefault(displayPageTemplateName);
-
-	const widgetId = getRandomString();
-
-	const widgetDefinition = getWidgetDefinition({
-		id: widgetId,
-		widgetName:
-			'com_liferay_asset_publisher_web_portlet_AssetPublisherPortlet',
-	});
-
-	const layout = await apiHelpers.headlessDelivery.createSitePage({
-		pageDefinition: getPageDefinition([widgetDefinition]),
-		siteId: site.id,
-		title: getRandomString(),
-	});
-
-	await pageEditorPage.goto(layout, site.friendlyUrlPath);
-
-	const topper = await pageEditorPage.getTopper(widgetId);
-
-	await topper.hover();
-	await clickAndExpectToBeVisible({
-		autoClick: true,
-		target: page.getByRole('menuitem', {
-			exact: true,
-			name: 'Configuration',
-		}),
-		trigger: topper.locator('.portlet-options'),
-	});
-
-	const assetPublisherConfigurationIframe = await page.frameLocator(
-		'iframe[title="Asset Publisher\\a      - Configuration"]'
-	);
-
-	const assetPublisherConfigurationDynamicRadio =
-		assetPublisherConfigurationIframe.getByText('Dynamic', {exact: true});
-	await assetPublisherConfigurationDynamicRadio.waitFor();
-	if (await assetPublisherConfigurationDynamicRadio.isHidden()) {
-		await assetPublisherConfigurationIframe
-			.getByRole('link', {name: 'Asset Selection'})
-			.click();
-	}
-	await assetPublisherConfigurationDynamicRadio.click();
-
-	const assetPublisherConfigurationSourceAssetTypeSelect =
-		await assetPublisherConfigurationIframe.getByLabel('Asset Type');
-	if (await assetPublisherConfigurationSourceAssetTypeSelect.isHidden()) {
-		await assetPublisherConfigurationIframe
-			.getByRole('link', {name: 'Source'})
-			.click();
-	}
-	await assetPublisherConfigurationSourceAssetTypeSelect.selectOption({
-		label: 'Blogs Entry',
-	});
-
-	await assetPublisherConfigurationIframe
-		.getByRole('button', {name: 'Save'})
-		.click();
-
-	await page.getByLabel('close', {exact: true}).click();
-	await page.getByLabel('Publish', {exact: true}).click();
-
-	await waitForSuccessAlert(
+	await friendlyURLCategoriesSetup({
+		apiHelpers,
+		displayPageTemplatesPage,
+		friendlyUrlCategories,
 		page,
-		'Success:The page was published successfully.'
-	);
+		pageEditorPage,
+		site,
+		vocabularyName,
+	});
 
 	await blogsEditBlogEntryPage.goto(site.friendlyUrlPath);
